EdgeScanner/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.40.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
ta>=0.11.0
